+++ src/app/(afterLogin)/[username]/_component/Nav.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useQuery } from "@tanstack/react-query";
+import { useQuery, useQueryClient } from "@tanstack/react-query";
 
 import getUserInfo from "../_lib/getUserInfo";
 
@@ -10,12 +10,14 @@ import { Navigation } from "./style";
 import { User } from "@/model/User";
 
 type Prop = {
-  username: any;
+  username: string;
 };
 
 export default function Nav({ username }: Prop) {
-  const { data } = useQuery<User, Object, User, [_1: string, _2: string]>({ queryKey: ["users", username], queryFn: getUserInfo });
-
+  const queryClient = useQueryClient();
+  const { data } = useQuery<User, Object, User, [_1: string, _2: string]>({ queryKey: ["user", username], queryFn: getUserInfo });
+  // const data = queryClient.getQueryData<User>(["user", username]);
+  // console.log("네비게이션 유저 데이터 : ", data);
   return (
     <Navigation>
       <BackBtn></BackBtn>
