+++ src/app/(afterLogin)/[username]/_component/ProfileUserData.tsx
@@ -1,21 +1,27 @@
 "use client";
 
-import { useQueryClient } from "@tanstack/react-query";
+import { useQuery, useQueryClient } from "@tanstack/react-query";
 import Link from "next/link";
-
+import { useSession } from "next-auth/react";
+import { Session } from "@auth/core/types";
 import Tab from "./Tab";
+import FollowButton from "../../_component/FollowButton";
 import { ProfileUserdataMid, UserName, SignupDate, AboutFollower, ProfileUserdata, NoAccountId, NoAccountMsg, ProfileUserdataTop, AbsoluteProfileContainer } from "./style";
 
 import { User } from "@/model/User";
+import getUserInfo from "../_lib/getUserInfo";
+
 type Prop = {
   username: string;
+  session: Session | null;
 };
 
-export default function ProfileUserData({ username }: Prop) {
+export default function ProfileUserData({ username, session }: Prop) {
   const queryClient = useQueryClient();
-  const data = queryClient.getQueryData<User>(["user", username]);
-
-  const createdDate = new Date(data?.createdAt!);
+  // 방문한 상세 페이지(내 정보 페이지 일 수도 있음)
+  const { data } = useQuery<User, Object, User, [_1: string, _2: string]>({ queryKey: ["user", username], queryFn: getUserInfo });
+  console.log("data in profileUserData", data);
+  // const createdDate = new Date(data?.createdAt!);
 
   if (data === undefined) {
     return (
@@ -25,6 +31,8 @@ export default function ProfileUserData({ username }: Prop) {
       </>
     );
   }
+
+  // return null;
   // 팔로우, 팔로잉 데이터 추가 필요
   return (
     <ProfileUserdata>
@@ -32,7 +40,9 @@ export default function ProfileUserData({ username }: Prop) {
         <AbsoluteProfileContainer>
           <div>아 몰랑</div>
         </AbsoluteProfileContainer>
-        <Link href="/settings/profile">프로필 수정</Link>
+        {/* 팔로우 버튼 확인 */}
+        {/* <Link href="/settings/profile">프로필 수정</Link> */}
+        {session?.user?.email === data?.id ? <Link href="/settings/profile">프로필 수정</Link> : <FollowButton user={data} session={session} />}
       </ProfileUserdataTop>
       <ProfileUserdataMid>
         <UserName>
@@ -45,14 +55,14 @@ export default function ProfileUserData({ username }: Prop) {
               <path d="M7 4V3h2v1h6V3h2v1h1.5C19.89 4 21 5.12 21 6.5v12c0 1.38-1.11 2.5-2.5 2.5h-13C4.12 21 3 19.88 3 18.5v-12C3 5.12 4.12 4 5.5 4H7zm0 2H5.5c-.27 0-.5.22-.5.5v12c0 .28.23.5.5.5h13c.28 0 .5-.22.5-.5v-12c0-.28-.22-.5-.5-.5H17v1h-2V6H9v1H7V6zm0 6h2v-2H7v2zm0 4h2v-2H7v2zm4-4h2v-2h-2v2zm0 4h2v-2h-2v2zm4-4h2v-2h-2v2z"></path>
             </g>
           </svg>
-          <div>{`가입일 ${createdDate.getFullYear()}년 ${createdDate.getMonth() + 1}월`}</div>
+          {/* <div>{`가입일 ${createdDate.getFullYear()}년 ${createdDate.getMonth() + 1}월`}</div> */}
         </SignupDate>
         <AboutFollower>
           <Link href="/follow">
-            <span>1</span> 팔로우 중
+            <span>{data._count.Followings}</span> 팔로우 중
           </Link>
           <Link href="/followers">
-            <span>0</span> 팔로워
+            <span>{data._count.Followers}</span> 팔로워
           </Link>
         </AboutFollower>
         <Tab></Tab>
