+++ src/app/(afterLogin)/[username]/_lib/getUserInfoServer.ts
@@ -1,13 +1,16 @@
 import { QueryFunction } from "@tanstack/query-core";
 import { User } from "@/model/User";
+import { cookies } from "next/headers";
 
-const getUserInfo: QueryFunction<User, [_1: string, _2: string]> = async ({ queryKey }) => {
+const getUserInfoServer: QueryFunction<User, [_1: string, _2: string]> = async ({ queryKey }) => {
   const [_1, username] = queryKey;
   const response = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/api/users/${username}`, {
-    credentials: "include",
+    headers: {
+      Cookie: cookies().toString(),
+    },
   });
 
   if (response.ok) return response.json();
 };
 
-export default getUserInfo;
+export default getUserInfoServer;
