+++ src/app/(afterLogin)/[username]/page.tsx
@@ -1,7 +1,8 @@
 import Link from "next/link";
 import { QueryClient, HydrationBoundary, dehydrate } from "@tanstack/react-query";
+import { auth } from "@/auth";
 
-import getUserInfo from "./_lib/getUserInfo";
+import getUserInfoServer from "./_lib/getUserInfoServer";
 import getUserPosts from "./_lib/getUserPosts";
 
 import Nav from "./_component/Nav";
@@ -16,11 +17,12 @@ type Props = {
 
 export default async function Username({ params }: Props) {
   const { username } = params;
-
+  console.log("username : ", username);
+  const session = await auth();
   const queryClient = new QueryClient();
   await queryClient.prefetchQuery({
     queryKey: ["user", username],
-    queryFn: getUserInfo,
+    queryFn: getUserInfoServer,
   });
   await queryClient.prefetchQuery({
     queryKey: ["posts", "user", username],
@@ -37,7 +39,7 @@ export default async function Username({ params }: Props) {
             <HeaderPhotoZone>
               <Link href="/home">{/* <Image src={이미지} alt="header_photo"></Image> */}</Link>
             </HeaderPhotoZone>
-            <ProfileUserData username={username} />
+            <ProfileUserData username={username} session={session} />
           </Profile>
           <UserPosts username={username} />
         </Userzone>
