+++ src/app/(afterLogin)/_component/Post.tsx
@@ -21,7 +21,9 @@ export default function Post({ post }: { post: IPost }) {
   const router = useRouter();
   const onClickArticle: MouseEventHandler<HTMLDivElement> = (e) => {
     e.stopPropagation();
-    router.push(`/${post.User.id}/status/${post.postId}`);
+    //
+    // router.push(`/${post.User.id}/status/${post.postId}`);
+    router.push(`/${post.User.id}}`);
   };
 
   // return null;
