+++ src/model/User.ts
@@ -1,12 +1,3 @@
-// { id: "elonmusk", nickname: "Elon Musk", image: faker.image.avatar(), UId: faker.number.int({ min: 10, max: 10000000000 }) },
-// export interface User {
-//   id: string;
-//   nickname: string;
-//   image: string;
-//   UId: number;
-//   createdAt: Date;
-// }
-
 interface UserID {
   id: string;
 }
@@ -14,7 +5,9 @@ interface UserID {
 export interface User {
   id: string;
   nickname: string;
+  // password: string;
   image: string;
+  // Posts: [];
   Followers: UserID[];
   _count: {
     Followers: number;
